Feat: guards를 활용한 예외 처리 추가
1. store 수정 권한 관련
2. store 삭제 권한 관련

+++ src/stores/stores.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { GetItemDto } from 'src/items/dto/get-item.dto';
 import { ItemsRepository } from 'src/items/items.repository';
 import { UpdateStoreReqDto } from './dto/store.request.dto';
@@ -33,6 +33,7 @@ export class StoresService {
 
   // * 가게 수정
   async updateStore(
+    userId: number,
     storeId: number,
     updateStoreDto: UpdateStoreReqDto,
   ): Promise<void> {
@@ -40,28 +41,28 @@ export class StoresService {
       storeId,
     );
     // ! 수정 권한이 없는 경우
-    // if () {
-    // throw new HttpException(
-    //   { message: '수정 권한이 없습니다.' },
-    //   HttpStatus.FORBIDDEN,
-    // );
-    // }
+    if (userId !== store.ownerId) {
+      throw new HttpException(
+        { message: '수정 권한이 없습니다.' },
+        HttpStatus.FORBIDDEN,
+      );
+    }
 
     return await this.storesRepository.updateStore(storeId, updateStoreDto);
   }
 
   // * 가게 삭제
-  async deleteStore(storeId: number): Promise<void> {
+  async deleteStore(userId: number, storeId: number): Promise<void> {
     const store: GetStoreResData = await this.storesRepository.selectOneStore(
       storeId,
     );
     // ! 삭제 권한이 없는 경우
-    // if () {
-    // throw new HttpException(
-    //   { message: '삭제 권한이 없습니다.' },
-    //   HttpStatus.FORBIDDEN,
-    // );
-    // }
+    if (userId !== store.ownerId) {
+      throw new HttpException(
+        { message: '삭제 권한이 없습니다.' },
+        HttpStatus.FORBIDDEN,
+      );
+    }
     return await this.storesRepository.deleteStore(storeId);
   }
 }
